Remove or replace Mozilla branding images and videos.
Replace the discord-bot-logo.png with a generic version.
Replace the hmc-logo.png with a generic version.
Remove the home-hero-background.jpg as it is unused.
Replace the spoke_logo_black.png with a generic version.
Remove discord.mp4 and discord.webm because they have too much branding to easily replace.
It would be nice to have replacements for the Discord videos eventually, for now they have been switched out on the Discord bot page with additional copies of the logo images.
Replace the HmcLogo.svg with a generic version.